Added ability to set the version number of the models
eg update_models.sh -v v1.15.90
Also added option to commit changes to git straight away

+++ scripts/update_models.sh
@@ -2,14 +2,25 @@
 
 set -e
 
-function get_aws_sdk_go
+usage()
+{
+  echo "Usage: update_models.sh -c [ -v MODELS_VERSION_NUMBER ]"
+  exit 2
+}
+
+get_aws_sdk_go()
 {
   # clone aws-sdk-go into folder
-  git clone --quiet --progress --depth=1 https://github.com/aws/aws-sdk-go.git $1
+  git clone https://github.com/aws/aws-sdk-go.git $1
+  if [ -n "$2" ]; then
+      pushd $1
+      git checkout $2
+      popd
+  fi
   return 0
 }
 
-function copy_model_files
+copy_model_files()
 {
     rm -rf $2/apis/*
     rm -rf $2/endpoints/*
@@ -17,10 +28,10 @@ function copy_model_files
     cp -R $1/endpoints/* $2/endpoints/
     rm $2/apis/*.go
     rm $2/endpoints/*.go
-  return 0
+    return 0
 }
 
-function build_files
+build_files()
 {
     # build the code generator and run it
     echo "Build the code generator"
@@ -32,12 +43,44 @@ function build_files
     swift build
 }
 
-function cleanup
+check_for_local_changes()
 {
-    rm -rf $TEMP_DIR
+    LOCAL_CHANGES=`git status --porcelain`
+    if [ -n "$LOCAL_CHANGES" ] && [ $COMMIT_CHANGES=1 ]; then
+        echo "You have local changes already and you have requested to commit changes after this script has run.\nEither remove the -c option or stash your local changes."
+        usage
+        exit
+    fi
 }
 
+commit_changes()
+{
+    COMMIT_MSG="Sync models with aws-sdk-go "$AWS_MODELS_VERSION
+    git commit -m "$COMMIT_MSG"
+}
+
+cleanup()
+{
+    if [ -n "$TEMP_DIR" ]; then
+        rm -rf $TEMP_DIR
+    fi
+}
+
+while getopts 'cv:' option
+do
+    case $option in
+        v) AWS_MODELS_VERSION=$OPTARG ;;
+        c) COMMIT_CHANGES=1 ;;
+        *)
+        usage
+        exit ;;
+    esac
+done
+
+
 trap cleanup EXIT
+
+check_for_local_changes
 
 #create temp folder
 TEMP_DIR=`mktemp -d`
@@ -46,7 +89,7 @@ echo "Using temp folder "$TEMP_DIR
 #get aws-sdk-go models
 echo "Get aws-sdk-go models"
 AWS_SDK_GO=$TEMP_DIR/aws-sdk-go/
-get_aws_sdk_go $AWS_SDK_GO
+get_aws_sdk_go $AWS_SDK_GO $AWS_MODELS_VERSION
 
 #copy aws-sdk-go models into aws-sdk-swift
 echo "Copy models to aws-sdk-swift"
@@ -56,3 +99,7 @@ copy_model_files $AWS_SDK_GO_MODELS $TARGET_MODELS
 
 #build service files from the models and check they compile
 build_files
+
+if [ $COMMIT_CHANGES=1 ]; then
+    commit_changes
+fi
